move get_fs_type() and find_non_nfs_dir() to file_util.py
