+++ mock/py/mockbuild/file_util.py
@@ -4,6 +4,7 @@ import errno
 import os
 import os.path
 import stat
+import subprocess
 import time
 
 from . import exception
@@ -96,3 +97,20 @@ def is_in_dir(path, directory):
     directory = os.path.realpath(directory)
 
     return os.path.commonprefix([path, directory]) == directory
+
+
+def get_fs_type(path):
+    cmd = ['/bin/stat', '-f', '-L', '-c', '%T', path]
+    p = subprocess.Popen(cmd, shell=False, stdout=subprocess.PIPE,
+                         universal_newlines=True)
+    p.wait()
+    with p.stdout as f:
+        return f.readline().strip()
+
+
+def find_non_nfs_dir():
+    dirs = ('/dev/shm', '/run', '/tmp', '/usr/tmp', '/')
+    for d in dirs:
+        if not get_fs_type(d).startswith('nfs'):
+            return d
+    raise exception.Error('Cannot find non-NFS directory in: %s' % dirs)
